+++ packages/proto-signing/src/index.ts
@@ -19,7 +19,7 @@ export type {
   TsProtoGeneratedType,
   TxBodyEncodeObject,
 } from "./registry";
-export { isPbjsGeneratedType, isTsProtoGeneratedType, isTxBodyEncodeObject, Registry } from "./registry";
+export { hasCreate, hasFromPartial, isTxBodyEncodeObject, Registry } from "./registry";
 export type { AccountData, Algo, DirectSignResponse, OfflineDirectSigner, OfflineSigner } from "./signer";
 export { isOfflineDirectSigner } from "./signer";
 export { makeAuthInfoBytes, makeSignBytes, makeSignDoc } from "./signing";